requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
discord.py==1.7.2
html5lib==1.1
lxml==4.6.3
requests==2.25.0